django/sample-app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via pip-tools
coverage==6.3.2
    # via -r requirements.in
cryptography==3.3.2
    # via -r requirements.in
decorator==5.1.1
    # via ipython
django==3.2.13
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-extensions
    #   silica-django
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-sass-processor==1.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
idna==3.3
    # via requests
ipython==8.2.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
parso==0.8.3
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.6.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via ipython
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.11.2
    # via ipython
pytz==2022.1
    # via django
requests==2.27.1
    # via -r requirements.in
silica-django==0.0.5a2
    # via -r requirements.in
six==1.16.0
    # via cryptography
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
tomli==2.0.1
    # via pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.1.1
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
